refactor(ai): extract usePromptFormState hook to achieve perfect code health
